refactor: just moved the internal structure of <CurricSEOAccordion/> around
