+++ src/components/CurriculumComponents/CurricSEOAccordion/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./CurricSEOAccordion";
