+++ src/components/CurriculumComponents/CurricVisualiserLayout/index.tsx
@@ -2,9 +2,8 @@ import { OakFlex } from "@oaknational/oak-components";
 import styled from "styled-components";
 import { PortableTextBlock } from "@portabletext/types";
 
-import { CurricSEOAccordion } from "../CurricSEOAccordion/CurricSEOAccordion";
-
 import { SubjectPhasePickerData } from "@/components/SharedComponents/SubjectPhasePicker/SubjectPhasePicker";
+import CurricSEOAccordion from "@/components/CurriculumComponents/CurricSEOAccordion";
 
 const CurriculumVisualiserLayoutLeft = styled(OakFlex)`
   min-width: 296px;
